backend/requirements.txt
fastapi
uvicorn
sqlalchemy
paho-mqtt
python-dotenv
psycopg2
pydantic-settings